Combine constants related to SymbolKind in one place
