+++ denops/@ddu-sources/lsp_workspaceSymbol.ts
@@ -5,7 +5,7 @@ import { Location, SymbolInformation, WorkspaceSymbol } from "npm:vscode-languag
 import { lspRequest, LspResult, Method } from "../ddu_source_lsp/request.ts";
 import { Client, ClientName, getClients } from "../ddu_source_lsp/client.ts";
 import { uriToPath } from "../ddu_source_lsp/util.ts";
-import { KindName } from "./lsp_documentSymbol.ts";
+import { KindName } from "../@ddu-filters/converter_lsp_symbol.ts";
 import { ActionData } from "../@ddu-kinds/lsp.ts";
 import { isValidItem } from "../ddu_source_lsp/handler.ts";
